Move get_unique_subjects function to utils/filemanip.py
