+++ clinica/pipelines/t1_freesurfer_longitudinal/t1_freesurfer_longitudinal_correction_utils.py
@@ -1,4 +1,5 @@
 # coding: utf8
+import clinica.utils.filemanip
 
 __author__ = "Alexis Guyot"
 __copyright__ = "Copyright 2016-2019, The Aramis Lab Team"
@@ -82,7 +83,7 @@ def check_reconalled(caps_dir, subject_list, session_list):
 
     # get unique list of subjects and corresponding per-subject sessions
     [unique_subject_list,
-     persubject_session_list2] = template_utils.get_unique_subjects(
+     persubject_session_list2] = clinica.utils.filemanip.get_unique_subjects(
          subject_list, session_list)
 
     # Check if t1-freesurfer-template has been correctly run
