+++ clinica/pipelines/t1_freesurfer_longitudinal/t1_freesurfer_template_utils.py
@@ -1,4 +1,5 @@
 # coding: utf8
+import clinica.utils.filemanip
 
 __author__ = "Alexis Guyot"
 __copyright__ = "Copyright 2016-2019, The Aramis Lab Team"
@@ -92,57 +93,6 @@ def check_xsectional_reconalled(caps_dir, subject_list, session_list):
                 error_msg = '{0} Re-run t1-freesurfer-cross-sectional.'.format(
                     error_msg)
                 raise IOError(error_msg)
-
-
-def get_unique_subjects(in_subject_list, in_session_list):
-    """Get unique participant IDs
-
-    The function to read the .tsv file returns the following
-    participant_id and session_id lists:
-    participant1, participant1, ..., participant2, participant2, ...
-    session1    , session2    , ..., session1    , session2    , ...
-    This function returns a list where all participants are only selected
-    once:
-    participant1, participant2, ..., participant_n
-    and for each participant, the list of corresponding session id
-    eg.:
-    participant1 -> [session1, session2]
-    participant2 -> [session1]
-    ...
-    participant_n -> [session1, session2, session3]
-
-    Args:
-        in_subject_list (list of strings): list of participant_id
-        in_session_list (list of strings): list of session_id
-
-    Returns:
-        out_unique_subject_list (list of strings): list of
-            participant_id, where each participant appears only once
-        out_persubject_session_list2 (list of list): list of list
-            (list2) of session_id associated to any single participant
-    """
-
-    import numpy as np
-
-    subject_array = np.array(in_subject_list)
-    session_array = np.array(in_session_list)
-
-    # The second returned element indicates for each participant_id the
-    # element they correspond to in the 'unique' list. We will use this
-    # to link each session_id in the repeated list of session_id to
-    # their corresponding unique participant_id
-
-    unique_subject_array, out_inverse_positions = np.unique(
-        subject_array, return_inverse=True)
-    out_unique_subject_list = unique_subject_array.tolist()
-
-    subject_number = len(out_unique_subject_list)
-    out_persubject_session_list2 = [
-        session_array[
-            out_inverse_positions == subject_index
-            ].tolist() for subject_index in range(subject_number)]
-
-    return out_unique_subject_list, out_persubject_session_list2
 
 
 def sessionid_to_sessionlabel(sessionid):
@@ -255,7 +205,7 @@ def get_longsubdir_dict(
     # list of associated sessions
     [
         unique_subject_list,
-        persubject_session_list2] = utils.get_unique_subjects(
+        persubject_session_list2] = clinica.utils.filemanip.get_unique_subjects(
             in_subject_list, in_session_list)
 
     # get longitudinal identifier for ech subject and store into
@@ -550,7 +500,7 @@ def check_caps_template(
 
     # get a unique list of subjects and associated per-subject list of
     # sessions
-    unique_subject_list, persubject_session_list2 = utils.get_unique_subjects(
+    unique_subject_list, persubject_session_list2 = clinica.utils.filemanip.get_unique_subjects(
         subject_list, session_list)
 
     # Check processed/unprocessed subjects
